+++ src/utils/convertVideoResource.test.ts
@@ -1,11 +1,11 @@
-import { streamVideoSample } from '../test-support/video-service-responses';
+import { VideoResourceFactory } from '../test-support/TestFactories';
 import convertPlaybackResource from './convertPlaybackResource';
 import convertVideoResource from './convertVideoResource';
 
 jest.mock('./convertPlaybackResource');
 
 it('maps the data to a Video', () => {
-  const videoResource = streamVideoSample;
+  const videoResource = VideoResourceFactory.streamSample();
 
   const video = convertVideoResource(videoResource);