+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Enhanced error handling to display a friendly error message
+
+### Fixed
+- BoclipsPlayer state reset before loading a new video 
+
 ## [1.0.1] - 2019-06-14
 
 ### Added